pyproject.toml
[project]
name = "projeto-nisa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.0.0",
    "dash-bootstrap-components>=2.0.4",
    "fastapi[standard]>=0.129.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
